+++ ctdcal/common.py
@@ -103,6 +103,7 @@ def validate_file(pathname, create=False):
         raise FileNotFoundError("The file %s could not be found." % str(p))
 
 
+# File imports
 def load_cnv(cnv_file: Union[str, Path]) -> pd.DataFrame:
     """
     Load Sea-Bird converted (.cnv) cast file into DataFrame
@@ -295,3 +296,40 @@ def load_exchange_ctd(
         comment="#",
         skipinitialspace=True,
     )
+
+
+# Utilities
+
+def print_progress_bar(
+        iteration,
+        total,
+        prefix="",
+        suffix="",
+        decimals=1,
+        length=100,
+        fill="█",
+        printEnd="\r",
+):
+    """
+    A progress bar, helpful for implementing into loops or highlighting progression through processing.
+
+    https://stackoverflow.com/questions/3173320/text-progress-bar-in-terminal-with-block-characters/13685020
+    credit: u/Greenstick
+    Call in a loop to create terminal progress bar
+    @params:
+        iteration   - Required  : current iteration (Int)
+        total       - Required  : total iterations (Int)
+        prefix      - Optional  : prefix string (Str)
+        suffix      - Optional  : suffix string (Str)
+        decimals    - Optional  : positive number of decimals in percent complete (Int)
+        length      - Optional  : character length of bar (Int)
+        fill        - Optional  : bar fill character (Str)
+        printEnd    - Optional  : end character (e.g. "\r", "\r\n") (Str)
+    """
+    percent = ("{0:." + str(decimals) + "f}").format(100 * (iteration / float(total)))
+    filledLength = int(length * iteration // total)
+    bar = fill * filledLength + "-" * (length - filledLength)
+    print(f"\r{prefix} |{bar}| {percent}% {suffix}", end=printEnd)  # Potential to add to log
+    # Print New Line on Complete
+    if iteration == total:
+        print()
